backend/requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12,<4.0
psycopg2>=2.8,<3.0
gunicorn>=20.0,<21.0
django-cors-headers>=3.5,<4.0
Pillow>=8.0,<9.0
requests>=2.25,<3.0
asgiref
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv